Combine Enterprise entries into a single text file
This matches the way the other entries are being set up and
makes it easy to strip out the references. The reason the
references are an issue is that any references that cite other
new entries won't resolve on the initial import into ADS and so
need to be submitted later.