Don't believe SG/6 with regard to SUN/2
All evidence suggests that RFWS misread the quick reference guide
in the 1994 era when determining what SUN MAG is associated with.

In the QRC the alphabetized section referred to the categories
and those categories included the document number. This meant that

    MAG   S2

on the alphabetical listing did not mean "SUN/2" but "section 2"
and looking at the Oct 1994 quick reference card (see Bulletin 14)
MAG does not have a SUN associated with it. Previously (Bulletin 9
April 1992) MAG is listed as APN/1.

In Bulletin 18 (March 1997) MAG is announced as a new product and
given SUN/171 so it's not clear that SG/6 was able to cite any SUN
at all in its 1994 vintage. For now I am deciding that SUN/2 was
a typo and to use SUN/171 for referencing. In theory SG/6 should
be fixed.